requirements.txt
flask
flask_sqlalchemy
psycopg2

gunicorn
python-dotenv
